% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimasim.R
\name{arimasim}
\alias{arimasim}
\title{Search for rigth seeds for the rigth ARIMA simulation with arima.sin() function using auto.arima() function}
\usage{
arimasim(
  a,
  z,
  n,
  ar11,
  ma11,
  ar22,
  ma22,
  ar33,
  ma33,
  p,
  d,
  q,
  sd = sd,
  j1,
  k1,
  j2,
  k2,
  j3,
  k3,
  arr1,
  maa1,
  arr2,
  maa2,
  arr3,
  maa3,
  n_cores
)
}
\arguments{
\item{a}{first seed boundary}

\item{z}{last seed boundary}

\item{n}{number of samples}

\item{ar11}{character to search for in third coefficient of autoregressive}

\item{ma11}{character to search for in third coefficient of autoregressive}

\item{ar22}{character to search for in third coefficient of autoregressive}

\item{ma22}{character to search for in third coefficient of autoregressive}

\item{ar33}{character to search for in third coefficient of autoregressive}

\item{ma33}{character to search for in third coefficient of autoregressive}

\item{p}{order of the autoregressive}

\item{d}{degree of difference}

\item{q}{degree of moving average}

\item{sd}{standard deviation of the series}

\item{j1}{length of character to search for in first coefficient of autoregressive}

\item{k1}{length of character to search for in third coefficient of autoregressive}

\item{j2}{length of character to search for in second coefficient of autoregressive}

\item{k2}{length of character to search for in third coefficient of autoregressive}

\item{j3}{length of character to search for in third coefficient of autoregressive}

\item{k3}{length of character to search for in third coefficient of autoregressive}

\item{arr1}{character to search for in first coefficient of autoregressive}

\item{maa1}{character to search for in third coefficient of autoregressive}

\item{arr2}{character to search for in second coefficient of autoregressive}

\item{maa2}{character to search for in third coefficient of autoregressive}

\item{arr3}{character to search for in third coefficient of autoregressive}

\item{maa3}{character to search for in third coefficient of autoregressive}

\item{n_cores}{number of core(s) to be used on your operaterating system}
}
\description{
More detailed Description
}
\section{Functions}{
\itemize{
\item \code{arimasim()}: arimasim helps to Search for rigth seeds for the rigth AR simulation with arima.sin() finction using auto.arima() function

}}
