% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimasim.R
\name{arimasim}
\alias{arimasim}
\title{Parameterized Simulation}
\usage{
arimasim(
  a,
  z,
  n,
  ar11,
  ma11,
  ar22,
  ma22,
  ar33,
  ma33,
  p,
  d,
  q,
  sd = sd,
  j1,
  k1,
  j2,
  k2,
  j3,
  k3,
  arr1,
  maa1,
  arr2,
  maa2,
  arr3,
  maa3
)
}
\arguments{
\item{a}{first seed boundary}

\item{z}{last seed boundary}

\item{n}{number of samples}

\item{ar11}{character to search for in third coefficient of autoregressive}

\item{ma11}{character to search for in third coefficient of autoregressive}

\item{ar22}{character to search for in third coefficient of autoregressive}

\item{ma22}{character to search for in third coefficient of autoregressive}

\item{ar33}{character to search for in third coefficient of autoregressive}

\item{ma33}{character to search for in third coefficient of autoregressive}

\item{p}{order of the autoregressive}

\item{d}{degree of difference}

\item{q}{degree of moving average}

\item{sd}{standard deviation of the series}

\item{j1}{length of character to search for in first coefficient of autoregressive}

\item{k1}{length of character to search for in third coefficient of autoregressive}

\item{j2}{length of character to search for in second coefficient of autoregressive}

\item{k2}{length of character to search for in third coefficient of autoregressive}

\item{j3}{length of character to search for in third coefficient of autoregressive}

\item{k3}{length of character to search for in third coefficient of autoregressive}

\item{arr1}{character to search for in first coefficient of autoregressive}

\item{maa1}{character to search for in third coefficient of autoregressive}

\item{arr2}{character to search for in second coefficient of autoregressive}

\item{maa2}{character to search for in third coefficient of autoregressive}

\item{arr3}{character to search for in third coefficient of autoregressive}

\item{maa3}{character to search for in third coefficient of autoregressive}
}
\value{
A data frame get printed to the console with its first colomn being the rank and the next few column could be the coefficients of AR or MA both with varying orders depending on the order and classes of ARIMA model being searched for. The last column of the data frame could be the intercept if any exist within the range of the search.
}
\description{
Parameterized Simulation
}
\section{Functions}{
\itemize{
\item \code{arimasim()}: arimasim helps to Search for rigth seeds for the rigth AR simulation with arima.sin() finction using auto.arima() function

Search for rigth seeds for the rigth ARIMA simulation with arima.sin() function using auto.arima() function

This function obtains a Random Number Generator (RNG) or collection of RNGs that replicate the required parameter(s) of a distribution for a time series of data. Consider the case of reproducing a time series data set of size 20 that uses an autoregressive (AR) model with phi = 0.8 and standard deviation equal to 1. When one checks the arima.sin() function's estimated parameters, it's possible that after a single trial or a few more, one won't find the precise parameters. This enables one to look for the ideal RNG setting for a simulation that will accurately duplicate the desired parameters.

}}
\examples{
  arimasim(a= 289800,z= 289989,n= 10,p= 1,d= 0,q= 0,ar11= 0.8,sd = 1,j1= 4,arr1= "0.80")

}
